pyproject.toml
[tool.poetry]
name = "appsec-discovery"
version = "0.7.1"
license = "MIT"
description = "Discover sensitive objects in project code"
authors = ["Dmitrii Mariushkin <d.v.marushkin@gmail.com>"]
readme = "README.md"
keywords = ["security", "assets", "discovery", "scanning", "appsec"]
packages = [{include = "appsec_discovery"}]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Security"
    
]

[tool.poetry.urls]
"Homepage" = "https://github.com/dmarushkin/appsec-discovery"
"Bug Tracker" = "https://github.com/dmarushkin/appsec-discovery/issues"

[tool.poetry.scripts]
appsec-discovery = "appsec_discovery.cli:main"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
semgrep = "1.62"
pydantic = "1.10.13"
pyyaml = "^6.0.2"
sarif-om = "^1.0.4"
jschema-to-python = "^1.2.3"
huggingface-hub = "^0.26.2"
llama-cpp-python = "^0.3.1"
mkdocs-material = "^9.5.44"
graphql-core = "^3.2.5"
proto-schema-parser = "^1.4.2"
openapi3-parser = "^1.1.19"
click-loglevel = "^0.5.0"
openai = "^1.60.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
